refactor: removes one directory level
