+++ utils/detect_arm.go
@@ -1,4 +1,4 @@
-package dsp
+package utils
 
 import (
 	"io"
